Refactor timesync to eliminate build error when not using time
