+++ components/zigbee/time/zigbee_time.h
@@ -2,7 +2,7 @@
 
 #include "esphome/core/component.h"
 #include "esphome/components/time/real_time_clock.h"
-#include "zigbee.h"
+#include "../zigbee.h"
 
 #define ZIGBEE_TIME_OFFSET 946684800 /* Zigbee time is based on counting seconds from 1 Jan 2000 (=946684800) */
 
@@ -17,8 +17,7 @@ class ZigbeeTime : public time::RealTimeClock {
     void setup() override;
     void loop() override;
     void update() override;
-    void recieve_timesync_response(esp_zb_zcl_read_attr_resp_variable_t *variable);
-    void send_timesync_request();
+    void sync(uint32_t utc);
 
   protected:
     ZigBeeComponent* zc_;
