+++ components/zigbee/zigbee.cpp
@@ -19,7 +19,6 @@ namespace esphome {
 namespace zigbee {
 
 ZigBeeComponent *zigbeeC;
-ZigbeeTime *zigbeeT;
 
 device_params_t coord;
 
@@ -224,10 +223,6 @@ void ZigBeeComponent::searchBindings() {
   esp_zb_zdo_binding_table_req(mb_req, bindingTableCb, (void *) mb_req);
 }
 
-void ZigBeeComponent::set_timeref(ZigbeeTime* zt) {
-  zigbeeT = zt;
-}
-
 static esp_err_t zb_attribute_handler(const esp_zb_zcl_set_attr_value_message_t *message) {
   esp_err_t ret = ESP_OK;
 
@@ -250,10 +245,10 @@ static esp_err_t zb_cmd_attribute_handler(const esp_zb_zcl_cmd_read_attr_resp_me
   switch(message->info.cluster) {
     case ESP_ZB_ZCL_CLUSTER_ID_TIME:
       ESP_LOGI(TAG, "Recieved time information");
-      if(zigbeeT==NULL) {
+      if(zigbeeC->timesync_callback_==NULL) {
         ESP_LOGI(TAG, "No time component linked to update time!");
       } else {
-        zigbeeT->recieve_timesync_response(message->variables);
+        zigbeeC->timesync_callback_(message->variables);
       }
       break;
     default:
@@ -400,7 +395,8 @@ static void esp_zb_task_(void *pvParameters) {
 
 void ZigBeeComponent::setup() {
   zigbeeC = this;
-  zigbeeT = NULL;
+  timesync_callback_ = NULL;
+
   esp_zb_platform_config_t config = {
       .radio_config = ESP_ZB_DEFAULT_RADIO_CONFIG(),
       .host_config = ESP_ZB_DEFAULT_HOST_CONFIG(),
